+++ tensorflow/lite/experimental/micro/micro_interpreter.cc
@@ -108,7 +108,7 @@ MicroInterpreter::MicroInterpreter(const Model* model,
 TfLiteStatus MicroInterpreter::AllocateInputAndActTensors() {
   const flatbuffers::Vector<flatbuffers::Offset<Buffer>>* buffers =
       model_->buffers();
-  for (int i = 0; i < subgraph_->inputs()->size(); ++i) {
+  for (size_t i = 0; i < subgraph_->inputs()->size(); ++i) {
     const int tensor_index = subgraph_->inputs()->Get(i);
     const auto* tensor = tensors_->Get(tensor_index);
     const TfLiteStatus status = tensor_allocator_->AllocateTensor(
@@ -123,20 +123,20 @@ TfLiteStatus MicroInterpreter::AllocateInputAndActTensors() {
       sizeof(int) * tensors_->size(), sizeof(int)));
   int* last_used = reinterpret_cast<int*>(tensor_allocator_->AllocateMemory(
       sizeof(int) * tensors_->size(), sizeof(int)));
-  for (int i = 0; i < tensors_->size(); ++i) {
+  for (size_t i = 0; i < tensors_->size(); ++i) {
     first_created[i] = -1;
     last_used[i] = -1;
   }
 
-  for (int i = (operators_->size() - 1); i >= 0; --i) {
+  for (size_t i = (operators_->size() - 1); i >= 0; --i) {
     const auto* op = operators_->Get(i);
-    for (int n = 0; n < op->inputs()->size(); ++n) {
+    for (size_t n = 0; n < op->inputs()->size(); ++n) {
       const int tensor_index = op->inputs()->Get(n);
       if ((last_used[tensor_index] == -1) || (last_used[tensor_index] < i)) {
         last_used[tensor_index] = i;
       }
     }
-    for (int n = 0; n < op->outputs()->size(); ++n) {
+    for (size_t n = 0; n < op->outputs()->size(); ++n) {
       const int tensor_index = op->outputs()->Get(n);
       const int create_before = i;
       int destroy_after = last_used[tensor_index];
@@ -156,7 +156,7 @@ TfLiteStatus MicroInterpreter::AllocateInputAndActTensors() {
     }
   }
 
-  for (int i = 0; i < tensors_->size(); ++i) {
+  for (size_t i = 0; i < tensors_->size(); ++i) {
     const auto* tensor = tensors_->Get(i);
     const bool is_read_only = (first_created[i] == -1) && (last_used[i] != -1);
     if (tensor->is_variable() || is_read_only) {
@@ -184,7 +184,7 @@ TfLiteStatus MicroInterpreter::Invoke() {
   }
   TfLiteStatus status = kTfLiteOk;
   auto opcodes = model_->operator_codes();
-  for (int i = 0; i < operators_->size(); ++i) {
+  for (size_t i = 0; i < operators_->size(); ++i) {
     const auto* op = operators_->Get(i);
     int index = op->opcode_index();
     if (index < 0 || index >= opcodes->size()) {
